Resolved warning

+++ FluxEngine/Audio/Sound.cpp
@@ -13,7 +13,7 @@ bool Sound::Load(InputStream& inputStream)
 	AUTOPROFILE_DESC(Sound_Load, inputStream.GetSource().c_str());
 	AudioEngine* pAudio = GetSubsystem<AudioEngine>();
 	FMOD_RESULT result = pAudio->GetSystem()->createSound(inputStream.GetSource().c_str(), FMOD_2D, nullptr, &m_pSound);
-	SetMemoryUsage(inputStream.GetSize());
+	SetMemoryUsage((uint32)inputStream.GetSize());
 	FLUX_LOG_FMOD(result);
 	return m_pSound != nullptr;
 }
